+++ src/components/NewDraftWizard/WizardHooks/WizardSaveDraftHook.tsx
@@ -4,7 +4,7 @@ import { resetDraftStatus } from "features/draftStatusSlice"
 import { setLoading, resetLoading } from "features/loadingSlice"
 import { updateStatus } from "features/statusMessageSlice"
 import { resetCurrentObject } from "features/wizardCurrentObjectSlice"
-import { addObjectToDrafts, replaceObjectInFolder } from "features/wizardSubmissionFolderSlice"
+import { addDraftObject, replaceObjectInFolder } from "features/wizardSubmissionFolderSlice"
 import draftAPIService from "services/draftAPI"
 import type { FolderDetailsWithId, ObjectDisplayValues } from "types"
 import { getObjectDisplayTitle } from "utils"
@@ -26,18 +26,10 @@ const saveDraftHook = async (props: SaveDraftHookProps) => {
 
   if (accessionId && objectStatus === ObjectStatus.draft) {
     const response = await draftAPIService.patchFromJSON(objectType, accessionId, values)
-    const index = folder.drafts.findIndex(item => item.accessionId === accessionId)
+
     if (response.ok) {
       dispatch(resetDraftStatus())
-      dispatch(
-        replaceObjectInFolder(
-          folder.folderId,
-          accessionId,
-          index,
-          { displayTitle: draftDisplayTitle },
-          ObjectStatus.draft
-        )
-      )
+      dispatch(replaceObjectInFolder(accessionId, { displayTitle: draftDisplayTitle }, ObjectStatus.draft))
       dispatch(
         updateStatus({
           status: ResponseStatus.success,
@@ -58,7 +50,7 @@ const saveDraftHook = async (props: SaveDraftHookProps) => {
     dispatch(resetLoading())
     return response
   } else {
-    const response = await draftAPIService.createFromJSON(objectType, values)
+    const response = await draftAPIService.createFromJSON(objectType, folder.folderId, values)
     if (response.ok) {
       dispatch(
         updateStatus({
@@ -69,7 +61,7 @@ const saveDraftHook = async (props: SaveDraftHookProps) => {
       )
       dispatch(resetDraftStatus())
       dispatch(
-        addObjectToDrafts(folder.folderId, {
+        addDraftObject({
           accessionId: response.data.accessionId,
           schema: "draft-" + objectType,
           tags: { displayTitle: draftDisplayTitle },